requirements.txt
pdfplumber==0.11.0
PyPDF2==3.0.1
pandas==2.2.0
pydantic==2.6.0
python-docx==1.1.0
pyyaml==6.0.1
pytest==8.0.0
python-dateutil==2.8.2
boto3==1.34.0
claude-agent-sdk==0.2.0
